// src/shared/api/notifications/types.ts
export type NotificationType =
    | 'info'
    | 'message'
    | 'alert'
    | 'payment-dorm'
    | 'payment-ed'
    | 'kpi-pps'
    | 'doc-for-review'
    | 'digital-services'
    | 'schedule'
    | 'version-update'
    | 'electronic-interaction'
    | 'allowance'

export type TNotification = {
    id: string
    title: string
    text: string
    type: NotificationType
    icon?: ChildrenType
    time?: string
    image?: string
    duration?: number
    goTo?: string
    date?: string
    onClick?: () => void
    onClose?: () => void
    canClose?: boolean
    pageId?: string
    fullText?: boolean
    loadingRemove?: boolean
}